% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testpackageTest-package.R
\docType{package}
\name{testpackageTest-package}
\alias{testpackageTest}
\alias{testpackageTest-package}
\title{testpackageTest: My Demo Package}
\description{
This is my basic package.
}
\author{
\strong{Maintainer}: Mune Satake \email{ys5340a@student.american.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
